… README.md
… src/eel_smart_engine.erl