This PR improves performance.

## Benchmarks

The measure was done using Benchee, an Elixir lib.

### Mix (mix.exs)

```elixir
defmodule EelBench.MixProject do
  # ...

  # Run "mix help deps" to learn about dependencies.
  defp deps do
    [
      {:benchee, "~> 1.1", only: :dev},
      {:eel, git: "https://github.com/williamthome/eel.git", branch: "feat/improve-performance"}
    ]
  end
end
```

### Script (benchmark.exs)

```elixir
list = Enum.to_list 1..10

Benchee.run(%{
  "eex" => fn ->
    EEx.eval_string "Foo<%= Enum.map(list, fn(x) -> %><p><%= x %></p><% end) %>Bar", [list: list]
  end,
  "eel" => fn ->
    :eel.eval("Foo<%= lists:map(fun(X) -> %><p><%= X .%></p><% end, List) .%>Bar", %{List: list})
  end
})
```

### Run

Get the deps

```
mix deps.get
```

and run the benchmark

```
mix run benchmark.exs
```

### Results

#### Machine config

```
Operating System: Linux
CPU Information: Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz
Number of Available Cores: 8
Available memory: 15.55 GB
Elixir 1.14.2
Erlang 25.3.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 14 s
```

#### Run 1

```
Name           ips        average  deviation         median         99th %
eel        10.13 K       98.67 μs    ±20.17%       94.70 μs      195.18 μs
eex         9.33 K      107.14 μs    ±16.44%      103.34 μs      193.03 μs

Comparison: 
eel        10.13 K
eex         9.33 K - 1.09x slower +8.47 μs
```

#### Run 2

```
Name           ips        average  deviation         median         99th %
eel         9.88 K      101.22 μs    ±19.09%       96.63 μs      194.44 μs
eex         9.34 K      107.03 μs    ±16.08%      102.91 μs      188.58 μs

Comparison: 
eel         9.88 K
eex         9.34 K - 1.06x slower +5.80 μs
```

#### Run 3

```
Name           ips        average  deviation         median         99th %
eel        10.00 K      100.01 μs    ±19.81%       96.08 μs      194.63 μs
eex         9.33 K      107.20 μs    ±16.84%      102.62 μs      191.34 μs

Comparison: 
eel        10.00 K
eex         9.33 K - 1.07x slower +7.19 μs
```